Remove osco_helper.py, consolidate into osco.py.
